space and property done

+++ src/Space.java
@@ -3,18 +3,18 @@
  */
 public class Space {
 	private String name;
-	private int type; //diferenciate luck spaces, property spaces and special spaces
+	private short type; //diferenciate luck spaces, property spaces and special spaces
     
-	public Space(String name, int type) {
-                this.name = name;
+	public Space(String name, short type) {
+        this.name = name;
 		this.type = type;
 	}
 
-        public String getName() {
-                return this.name;
-        }
+    public String getName() {
+        return this.name;
+    }
 
-        public int getType() {
-                return this.type;
-        }
+    public short getType() {
+        return this.type;
+    }
 }
